# components/manager/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Fetcher Manager component",
        "title": "Fetcher Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:4006",
    "basePath": "/",
    "paths": {
        "/v1/fetcher": {
            "post": {
                "description": "Create a new data extraction job. The request will be validated, deduplicated within a 5-minute window, and all referenced connections will be tested before job creation. The metadata.source field is required for product isolation and datasource ownership validation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fetcher"
                ],
                "summary": "Create fetcher job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Fetcher request payload. metadata.source is required.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.FetcherRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate request - returning existing job",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.FetcherResponse"
                        }
                    },
                    "202": {
                        "description": "Job created and queued for processing",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.FetcherResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/fetcher/{id}": {
            "get": {
                "description": "Retrieve detailed information about a specific data extraction job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fetcher"
                ],
                "summary": "Get fetcher job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.JobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_LerianStudio_fetcher_pkg_model.DataRequest": {
            "description": "DataRequest encapsulates field mappings and optional filters for data extraction.",
            "type": "object",
            "required": [
                "mappedFields"
            ],
            "properties": {
                "filters": {
                    "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.NestedFilters"
                },
                "mappedFields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.FetcherRequest": {
            "description": "FetcherRequest represents the request body for creating a new data extraction job.",
            "type": "object",
            "required": [
                "dataRequest"
            ],
            "properties": {
                "dataRequest": {
                    "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.DataRequest"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.FetcherResponse": {
            "description": "FetcherResponse represents the response after successfully creating a data extraction job.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.JobResponse": {
            "description": "JobResponse represents the complete information about a data extraction job.",
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "filters": {
                    "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.NestedFilters"
                },
                "id": {
                    "type": "string"
                },
                "mappedFields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "organizationId": {
                    "type": "string"
                },
                "requestHash": {
                    "type": "string"
                },
                "resultHmac": {
                    "type": "string"
                },
                "resultPath": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.NestedFilters": {
            "type": "object",
            "additionalProperties": {
                "type": "object",
                "additionalProperties": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model_job.FilterCondition"
                    }
                }
            }
        },
...
        "github_com_LerianStudio_fetcher_pkg_model_job.FilterCondition": {
            "type": "object",
            "properties": {
                "between": {
                    "description": "Between specifies a range condition with exactly two values [min, max].\nMatches records where min \u003c= field \u003c= max\nExample: {\"between\": [100, 1000]} matches records where 100 \u003c= field \u003c= 1000",
                    "type": "array",
                    "items": {}
                },
                "eq": {
                    "description": "Equals specifies exact value matches. Multiple values treated as OR conditions.\nExample: {\"eq\": [\"active\", \"pending\"]} matches records where field equals \"active\" OR \"pending\"",
                    "type": "array",
                    "items": {}
                },
                "gt": {
                    "description": "GreaterThan specifies values that must be greater than the provided value.\nShould contain exactly one value for comparison.\nExample: {\"gt\": [100]} matches records where field \u003e 100",
                    "type": "array",
                    "items": {}
                },
                "gte": {
                    "description": "GreaterOrEqual specifies values that must be greater than or equal to the provided value.\nShould contain exactly one value for comparison.\nExample: {\"gte\": [\"2025-06-01\"]} matches records where field \u003e= \"2025-06-01\"",
                    "type": "array",
                    "items": {}
                },
                "in": {
                    "description": "In specifies a list of values where the field must match any one of them.\nMultiple values treated as OR conditions.\nExample: {\"in\": [\"active\", \"pending\", \"suspended\"]} matches any of these statuses",
                    "type": "array",
                    "items": {}
                },
                "like": {
                    "description": "Like specifies values that must match the provided value using LIKE pattern matching.\nShould contain exactly one value for comparison.\nExample: {\"like\": [\"%active%\"]} matches any status containing \"active\"",
                    "type": "array",
                    "items": {}
                },
                "lt": {
                    "description": "LessThan specifies values that must be less than the provided value.\nShould contain exactly one value for comparison.\nExample: {\"lt\": [1000]} matches records where field \u003c 1000",
                    "type": "array",
                    "items": {}
                },
                "lte": {
                    "description": "LessOrEqual specifies values that must be less than or equal to the provided value.\nShould contain exactly one value for comparison.\nExample: {\"lte\": [\"2025-06-30\"]} matches records where field \u003c= \"2025-06-30\"",
                    "type": "array",
                    "items": {}
                },
                "ne": {
                    "description": "NotEquals specifies values that must NOT match the provided value.\nShould contain exactly one value for comparison.\nExample: {\"ne\": [\"active\"]} excludes this status",
                    "type": "array",
                    "items": {}
                },
                "nin": {
                    "description": "NotIn specifies a list of values where the field must NOT match any of them.\nMultiple values treated as AND NOT conditions.\nExample: {\"nin\": [\"deleted\", \"archived\"]} excludes these statuses",
                    "type": "array",
                    "items": {}
                }
            }
        },
        "pkg.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}